Temp/requirements.txt
# === Core Web Server ===
flask==3.0.3
flask-cors==4.0.0
gunicorn==22.0.0

# === OpenAI + Embedding Pipeline ===
openai>=1.12.0
tiktoken>=0.7.0
faiss-cpu>=1.8.0
numpy>=1.26.0
pandas>=2.2.0

# === Document Parsing ===
pypdf>=4.2.0
python-docx>=1.1.2

# === Utilities ===
python-dotenv>=1.0.1
urllib3>=2.0.7
